Support schema_list instead of inheritance
Previously it was the model to create a model class by
inheriting from all the base schemas required, e.g.

class person(rfc.person, rfc.organizationalPerson, rfc.inetOrgPerson):
    ...

However this requires use of multiple inheritance which is a
complication that is not required.

Now we support doing:

class person(tldap.base.LDAPobject):
    schema_list = [rfc.person, rfc.organizationalPerson, rfc.inetOrgPerson]
    ...

Which is a lot simpler as we end up having only one parent.

This in turn means that super(...) will call the parent class as
expected and is easier to understand what is going on.

The previous technique will still work for now.