Remove using= parameter from object methods.
This means if we load object from LDAP server we have to save it to that
same LDAP server.

In a practical sense, this was the case anyway.

Trying to make things too generic just overcomplicates everything.

This also means that the dn for an object will not change, unless rename
is called. To try and do otherwise is just so very very confusing.