Delete using parameter.
Was required before 9f79c500bc971ad7fdd3dd6e4eb45d1b187d5f03, as we
couldn't be certain what LDAP server would be used.

In practise it was a complication, got me totally confused, and not sure
if I ever implemented it correctly.

Now we can just use self._alias, as this is guaranteed not to change for
the lifetime of the object.