Python3 fixes in dn parsing.
We use to take a UTF-8 encoded string, and compare byte by byte
for compliance with the standards.

Now we take a unicode string, and compare character by character
for compliance with the standards.

This means we can't check each byte in a multibyte character UTF_8
character sequence, so we don't even try.

Change-Id: I0a85279876063c1547d5f09ac1906261ee07e685